WAN 2.2 S2V: split CausalConv1d + MotionEncoder_tc into auxi_blocks.py
Our audio_utils.py was a kitchen-sink holding classes from three of the
reference's s2v module files. The most obvious miss: the reference has
a dedicated auxi_blocks.py for the low-level conv primitives, and we
didn't.

Move ``CausalConv1d`` and ``MotionEncoder_tc`` from audio_utils.py into
a new auxi_blocks.py, mirroring wan/modules/s2v/auxi_blocks.py exactly.

After the split, each file's purpose matches the reference:

  auxi_blocks.py  — CausalConv1d, MotionEncoder_tc
                    (ref wan/modules/s2v/auxi_blocks.py)
  audio_utils.py  — CausalAudioEncoder, AdaLayerNormZero, AudioInjector_WAN
                    (ref wan/modules/s2v/audio_utils.py minus
                    AudioCrossAttention — the cached_kv slot on
                    WanAttention covers that role)

audio_utils.py shrunk 559 → 221 lines. Imports trimmed: dropped the
unused ``Linear`` and ``get_conv3d_config`` references; added
``from .auxi_blocks import MotionEncoder_tc``.

No behavioral change. 56 tests collect cleanly.

AdaLayerNormZero stays in audio_utils.py (vs ref's model_s2v.py) —
co-located with AudioInjector_WAN since both are part of the audio-
conditioning path, cleaner than the ref's "shove it into the main model
file" pattern.
